requirements.txt
fastapi==0.116.1
uvicorn
openai==1.99.9
python-dotenv
pinecone==7.3.0
google-api-python-client
streamlit